Break out Enigma superclass responsibilities to Encryptor, Decryptor, Cracker subclasses, add tests to correspond to new classes
