--- runner.rb
@@ -1,4 +1,0 @@
-require './lib/enigma'
-
-enigma = Enigma.new
-require 'pry'; binding.pry
